+++ packages/nns/candid/governance.d.ts
@@ -126,6 +126,9 @@ export interface Committed {
 export interface Configure {
   operation: [] | [Operation];
 }
+export interface Countries {
+  iso_codes: Array<string>;
+}
 export interface CreateServiceNervousSystem {
   url: [] | [string];
   governance_parameters: [] | [GovernanceParameters];
@@ -174,6 +177,9 @@ export interface Follow {
 }
 export interface Followees {
   followees: Array<NeuronId>;
+}
+export interface GlobalTimeOfDay {
+  seconds_after_utc_midnight: [] | [bigint];
 }
 export interface Governance {
   default_followees: Array<[number, Followees]>;
@@ -275,6 +281,7 @@ export interface ListProposalInfo {
   before_proposal: [] | [NeuronId];
   limit: number;
   exclude_topic: Int32Array;
+  include_all_manage_neuron_proposals: [] | [boolean];
   include_status: Int32Array;
 }
 export interface ListProposalInfoResponse {
@@ -565,13 +572,18 @@ export interface SwapDistribution {
 }
 export interface SwapParameters {
   minimum_participants: [] | [bigint];
+  duration: [] | [Duration];
   neuron_basket_construction_parameters:
     | []
     | [NeuronBasketConstructionParameters];
+  confirmation_text: [] | [string];
   maximum_participant_icp: [] | [Tokens];
   minimum_icp: [] | [Tokens];
   minimum_participant_icp: [] | [Tokens];
+  start_time: [] | [GlobalTimeOfDay];
   maximum_icp: [] | [Tokens];
+  neurons_fund_investment_icp: [] | [Tokens];
+  restricted_countries: [] | [Countries];
 }
 export interface Tally {
   no: bigint;