+++ packages/nns/src/types/governance_converters.ts
@@ -155,6 +155,11 @@ export interface ListProposalsRequest {
   // seeing
   excludeTopic: Array<Topic>;
 
+  // Include all ManageNeuron proposals regardless of the visibility of the
+  // proposal to the caller principal. Note that exclude_topic is still
+  // respected even when this option is set to true.
+  includeAllManageNeuronProposals: boolean;
+
   // Include proposals that have a status in this list (see
   // [ProposalStatus] for more information). If this list is empty, no
   // restriction is applied.
@@ -497,6 +502,14 @@ export interface Duration {
   seconds?: bigint;
 }
 
+export interface GlobalTimeOfDay {
+  secondsAfterUtcMidnight?: bigint;
+}
+
+export interface Countries {
+  isoCodes: Array<string>;
+}
+
 export interface Tokens {
   e8s?: bigint;
 }
@@ -537,11 +550,16 @@ export interface NeuronBasketConstructionParameters {
 }
 export interface SwapParameters {
   minimumParticipants?: bigint;
+  duration?: Duration;
   neuronBasketConstructionParameters?: NeuronBasketConstructionParameters;
+  confirmationText?: string;
   maximumParticipantIcp?: Tokens;
+  neuronsFundInvestmentIcp?: Tokens;
   minimumIcp?: Tokens;
   minimumParticipantIcp?: Tokens;
+  startTime?: GlobalTimeOfDay;
   maximumIcp?: Tokens;
+  restrictedCountries?: Countries;
 }
 
 export interface SwapDistribution {