+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-# 0.18.1 (wip)
+# 0.18.2 (wip)
+
+## Release
+
+- nns `v0.16.5`
+- sns `v0.0.20`
+- cmc `v0.0.16`
+- ledger `v0.0.13`
+- ckBTC `v0.0.9`
+- ic-management `v0.0.6`
+- utils `v0.0.20`
+- nns-proto `v0.0.6`
+
+## Features
+
+- Reviver and replacer to interpret `BigInt`, `Principal`, and `Uint8Array` in `JSON.stringify|parse`
+
+# 0.18.1 (2023-08-07)
 
 ## Release
 
@@ -10,6 +27,20 @@
 - ic-management `v0.0.5`
 - utils `v0.0.19`
 - nns-proto `v0.0.5`
+
+## Features
+
+- update `nns/governance.did` and `sns_wasm.did` for 1-proposal
+- update `ic.did` for `@dfinity/ic-management`
+- update `icrc1_ledger.did`
+
+## Build
+
+- bump agent-js `v0.18.1`
+
+## Docs
+
+- add mention that ICRC-22 is currently a draft
 
 # 0.18.0 (2023-07-03)